move scripts to separate folder
