+++ .github/workflows/sync-knowledge-base.yml
@@ -1,13 +1,21 @@
-name: Trigger docs → OpenAI sync
+name: Sync knowledge-base docs to OpenAI vector-store
 
 on:
   push:
     branches: [master, rag]
-    paths: ['rag/knowledge-base/**']     # run only when docs change
-  workflow_dispatch:                # allow manual “Run workflow”
+    paths: ['**/**']     # run only when docs change
+  workflow_dispatch:            # allow manual “Run workflow” in UI
 
 jobs:
-  call-sync:
-    uses: ./rag/.ci/sync-knowledge-base.yml
-    secrets:
-      OPENAI_API_KEY: ${{ secrets.OPENAI_API_KEY }}
+  sync:
+    runs-on: ubuntu-latest
+
+    steps:
+      - uses: actions/checkout@v4
+
+      - name: Upload / prune Markdown files
+        uses: shmatt/assistants-vector-store-sync@v1
+        with:
+          pattern: 'rag/knowledge_base/**/*.md'              # glob relative to repo root
+          token:   ${{ secrets.OPENAI_API_KEY }}
+          key:     emailsmart-knowledge-base             # becomes the vector-store name
