+++ rag/scripts/db.ts
@@ -1,5 +1,5 @@
 import pg from 'pg';
-import { PG_URL } from './env.js';
+import { PG_URL } from '../src/env.js';
 
 // Create the database pool
 const db = new pg.Pool({ connectionString: PG_URL });
